requirements.txt
pytest>=7.0.0
allure-pytest>=2.9.0
selenium>=4.0.0
webdriver-manager>=3.8.0
allure-python-commons~=2.13.2